+++ My Solutions/079_Twice as old/src/TwiceAsOld.java
@@ -1,5 +1,5 @@
 public class TwiceAsOld {
-    public static int TwiceAsOld(int dadYears, int sonYears) throws Exception {
+    public static int TwiceAsOld(int dadYears, int sonYears){
         int ageDifference = dadYears - sonYears;
         int targetAge = sonYears * 2;
         int yearsDifference = Math.abs(targetAge - dadYears);
